Подмножества Stream API: distinct, limit, skip Основы Stream API

+++ src/Product.java
@@ -1,8 +1,8 @@
 class Product {
     String name;
-    double price;
+    int price;
 
-    Product(String name, double price) {
+    Product(String name, int price) {
         this.name = name;
         this.price = price;
     }
